+++ CHANGELOG.md
@@ -15,6 +15,11 @@ This Changelog tracks changes to this project. The notes below include a summary
   - to prevent future 412 error on account reactivation
 - `changed` schemas bump for shipping manifest requirement relaxing
 
+## 13 Jun 2022
+
+- `changed` compression for downloaded batches from gztar to zip
+- `changed` max download size from 100MB to 1GiB
+
 ## Version `0.26.14` - 9 Jun 2022
 
 - `changed` schemas bump for Microbiome support